+++ web/frontend/src/components/camera/camera.svelte
@@ -2,23 +2,18 @@
 
 import { onMount, onDestroy } from 'svelte';
 import { displayError } from '@/lib/error';
-import {
-  CameraClient,
-  Client,
-  type ResponseStream,
-  robotApi,
-  type ServiceError,
-} from '@viamrobotics/sdk';
+import { CameraClient, type ServiceError } from '@viamrobotics/sdk';
 import { selectedMap } from '@/lib/camera-state';
 import type { StreamManager } from './stream-manager';
+import { useClient } from '@/hooks/use-client';
 
 export let cameraName: string;
-export let client: Client;
 export let showExportScreenshot: boolean;
 export let refreshRate: string | undefined;
 export let streamManager: StreamManager;
-export let statusStream: ResponseStream<robotApi.StreamStatusResponse> | null;
 export let triggerRefresh = false;
+
+const { client, statusStream } = useClient();
 
 let imgEl: HTMLImageElement;
 let videoEl: HTMLVideoElement;
@@ -51,7 +46,7 @@ const updateCameraRefreshRate = () => {
 const exportScreenshot = async () => {
   let blob;
   try {
-    blob = await new CameraClient(client, cameraName).renderFrame(
+    blob = await new CameraClient($client, cameraName).renderFrame(
       'image/jpeg'
     );
   } catch (error) {
@@ -63,7 +58,7 @@ const exportScreenshot = async () => {
 };
 
 onMount(() => {
-  statusStream?.on('end', () => clearFrameInterval());
+  $statusStream?.on('end', () => clearFrameInterval());
 
   videoEl.srcObject = cameraManager.videoStream;
