+++ web/frontend/src/components/gamepad/index.svelte
@@ -4,20 +4,15 @@ import { grpc } from '@improbable-eng/grpc-web';
 import { onMount, onDestroy } from 'svelte';
 import { Timestamp } from 'google-protobuf/google/protobuf/timestamp_pb';
 import { ConnectionClosedError } from '@viamrobotics/rpc';
-import {
-  Client,
-  inputControllerApi as InputController,
-  type ResponseStream,
-  robotApi,
-  type ServiceError,
-} from '@viamrobotics/sdk';
+import { inputControllerApi as InputController, type ServiceError } from '@viamrobotics/sdk';
 import { notify } from '@viamrobotics/prime';
 import { rcLogConditionally } from '@/lib/log';
-import Collapse from '@/components/collapse.svelte';
+import Collapse from '@/lib/components/collapse.svelte';
+import { useClient } from '@/hooks/use-client';
 
 export let name: string;
-export let client: Client;
-export let statusStream: ResponseStream<robotApi.StreamStatusResponse> | null;
+
+const { client, statusStream } = useClient()
 
 let gamepadIdx: number | null = null;
 let gamepadConnectedPrev = false;
@@ -57,7 +52,7 @@ const sendEvent = (newEvent: InputController.Event) => {
   req.setController(name);
   req.setEvent(newEvent);
   rcLogConditionally(req);
-  client.inputControllerService.triggerEvent(req, new grpc.Metadata(), (error: ServiceError | null) => {
+  $client.inputControllerService.triggerEvent(req, new grpc.Metadata(), (error: ServiceError | null) => {
     if (error) {
       if (ConnectionClosedError.isError(error)) {
         return;
@@ -255,7 +250,7 @@ onMount(() => {
     return;
   }
   prevStates = { ...prevStates, ...curStates };
-  statusStream?.on('end', () => clearTimeout(handle));
+  $statusStream?.on('end', () => clearTimeout(handle));
   tick();
 });
