+++ web/frontend/src/components/gantry/index.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
 
 import { grpc } from '@improbable-eng/grpc-web';
-import { Client, gantryApi } from '@viamrobotics/sdk';
+import { gantryApi } from '@viamrobotics/sdk';
 import { displayError } from '@/lib/error';
 import { rcLogConditionally } from '@/lib/log';
-import Collapse from '@/components/collapse.svelte';
+import Collapse from '@/lib/components/collapse.svelte';
+import { useClient } from '@/hooks/use-client';
 
 export let name: string;
-export let client: Client;
 export let status: {
   is_moving: boolean
   lengths_mm: number[]
@@ -17,6 +17,8 @@ export let status: {
   lengths_mm: [],
   positions_mm: [],
 };
+
+const { client } = useClient();
 
 $: parts = status.lengths_mm.map((_, index) => ({
   axis: index,
@@ -46,7 +48,7 @@ const increment = (axis: number, amount: number) => {
   req.setPositionsMmList(pos);
 
   rcLogConditionally(req);
-  client.gantryService.moveToPosition(req, new grpc.Metadata(), displayError);
+  $client.gantryService.moveToPosition(req, new grpc.Metadata(), displayError);
 };
 
 const stop = () => {
@@ -54,7 +56,7 @@ const stop = () => {
   req.setName(name);
 
   rcLogConditionally(req);
-  client.gantryService.stop(req, new grpc.Metadata(), displayError);
+  $client.gantryService.stop(req, new grpc.Metadata(), displayError);
 };
 
 </script>
