+++ web/frontend/src/components/gripper/index.svelte
@@ -1,20 +1,22 @@
 <script lang="ts">
 
 import { grpc } from '@improbable-eng/grpc-web';
-import { Client, gripperApi } from '@viamrobotics/sdk';
+import { gripperApi } from '@viamrobotics/sdk';
 import { displayError } from '../../lib/error';
 import { rcLogConditionally } from '../../lib/log';
-import Collapse from '@/components/collapse.svelte';
+import Collapse from '@/lib/components/collapse.svelte';
+import { useClient } from '@/hooks/use-client';
 
 export let name: string;
-export let client: Client;
+
+const { client } = useClient();
 
 const stop = () => {
   const request = new gripperApi.StopRequest();
   request.setName(name);
 
   rcLogConditionally(request);
-  client.gripperService.stop(request, new grpc.Metadata(), displayError);
+  $client.gripperService.stop(request, new grpc.Metadata(), displayError);
 };
 
 const open = () => {
@@ -22,7 +24,7 @@ const open = () => {
   request.setName(name);
 
   rcLogConditionally(request);
-  client.gripperService.open(request, new grpc.Metadata(), displayError);
+  $client.gripperService.open(request, new grpc.Metadata(), displayError);
 };
 
 const grab = () => {
@@ -30,7 +32,7 @@ const grab = () => {
   request.setName(name);
 
   rcLogConditionally(request);
-  client.gripperService.grab(request, new grpc.Metadata(), displayError);
+  $client.gripperService.grab(request, new grpc.Metadata(), displayError);
 };
 
 </script>
