+++ web/frontend/src/components/motor/index.svelte
@@ -3,23 +3,25 @@
 import { Client, motorApi, MotorClient, type ServiceError } from '@viamrobotics/sdk';
 import { displayError } from '@/lib/error';
 import { rcLogConditionally } from '@/lib/log';
-import Collapse from '../collapse.svelte';
-
-const motorPosFormat = new Intl.NumberFormat(undefined, {
-  maximumFractionDigits: 3,
-});
+import Collapse from '../../lib/components/collapse.svelte';
+import { useClient } from '@/hooks/use-client';
 
 export let name: string;
-export let client: Client;
 export let status: undefined | {
   is_powered?: boolean
   position?: number
   is_moving?: boolean
 };
 
+const motorPosFormat = new Intl.NumberFormat(undefined, {
+  maximumFractionDigits: 3,
+});
+
+const { client } = useClient();
+
 type MovementTypes = 'go' | 'goFor' | 'goTo';
 
-const motorClient = new MotorClient(client, name, {
+const motorClient = new MotorClient($client, name, {
   requestLogger: rcLogConditionally,
 });
