+++ web/frontend/src/components/operations-sessions/index.svelte
@@ -1,23 +1,22 @@
 <script lang="ts">
 
 import { grpc } from '@improbable-eng/grpc-web';
-import { Client, robotApi } from '@viamrobotics/sdk';
+import { robotApi } from '@viamrobotics/sdk';
 import { displayError } from '@/lib/error';
 import { rcLogConditionally } from '@/lib/log';
-import Collapse from '../collapse.svelte';
+import Collapse from '../../lib/components/collapse.svelte';
+import { useClient } from '@/hooks/use-client';
 
-export let operations: { op: robotApi.Operation.AsObject; elapsed: number }[];
-export let sessions: robotApi.Session.AsObject[];
 export let sessionsSupported: boolean;
-export let connectionManager: { rtt: number };
-export let client: Client;
+
+const { client, operations, sessions, rtt } = useClient()
 
 const killOperation = (id: string) => {
   const req = new robotApi.CancelOperationRequest();
   req.setId(id);
 
   rcLogConditionally(req);
-  client.robotService.cancelOperation(req, new grpc.Metadata(), displayError);
+  $client.robotService.cancelOperation(req, new grpc.Metadata(), displayError);
 };
 
 const peerConnectionType = (info?: robotApi.PeerConnectionInfo.AsObject) => {
@@ -44,20 +43,20 @@ const peerConnectionType = (info?: robotApi.PeerConnectionInfo.AsObject) => {
   <div class="border border-t-0 border-medium p-4 text-xs">
     <div class="mb-4 flex gap-2 justify-end items-center">
       <label>RTT:</label>
-      {#if connectionManager.rtt < 50}
+      {#if rtt.current < 50}
         <v-badge
           variant="green"
-          label={`${connectionManager.rtt} ms`}
+          label={`${$rtt} ms`}
         />
-      {:else if connectionManager.rtt < 500}
+      {:else if rtt.current < 500}
         <v-badge
           variant="orange"
-          label={`${connectionManager.rtt} ms`}
+          label={`${rtt.current} ms`}
         />
       {:else}
         <v-badge
           variant="red"
-          label={`${connectionManager.rtt} ms`}
+          label={`${rtt.current} ms`}
         />
       {/if}
     </div>
@@ -72,11 +71,11 @@ const peerConnectionType = (info?: robotApi.PeerConnectionInfo.AsObject) => {
           <th class="border border-medium p-2">elapsed time</th>
           <th class="border border-medium p-2" />
         </tr>
-        {#each operations as { op, elapsed } (op.id)}
+        {#each $operations as { op, elapsed } (op.id)}
           <tr>
             <td class="border border-medium p-2">
               {op.id}
-              {#if client.sessionId === op.sessionId}
+              {#if $client.sessionId === op.sessionId}
                 <span class="font-bold">(this session)</span>
               {/if}
             </td>
@@ -101,11 +100,11 @@ const peerConnectionType = (info?: robotApi.PeerConnectionInfo.AsObject) => {
             <th class="border border-medium p-2">remote address</th>
             <th class="border border-medium p-2">local address</th>
           </tr>
-          {#each sessions as session (session.id)}
+          {#each $sessions as session (session.id)}
             <tr>
               <td class="border border-medium p-2">
                 {session.id}
-                {#if client.sessionId && session.id === client.sessionId}
+                {#if session.id === $client.sessionId}
                   <span class="font-bold">(ours)</span>
                 {/if}
               </td>
