+++ web/frontend/src/components/sensors/index.svelte
@@ -9,7 +9,8 @@
   import { notify } from '@viamrobotics/prime';
   import { resourceNameToString } from '@/lib/resource';
   import { rcLogConditionally } from '@/lib/log';
-  import Collapse from '../collapse.svelte';
+  import Collapse from '../../lib/components/collapse.svelte';
+  import { useClient } from '@/hooks/use-client';
 
   interface SensorName {
     name: string;
@@ -20,7 +21,8 @@
 
   export let name: string;
   export let sensorNames: SensorName[];
-  export let client: Client;
+
+  const { client } = useClient();
 
   interface Reading {
     _type: string;
@@ -44,7 +46,7 @@
     req.setSensorNamesList(names);
 
     rcLogConditionally(req);
-    client.sensorsService.getReadings(
+    $client.sensorsService.getReadings(
       req,
       new grpc.Metadata(),
       (
