+++ web/frontend/src/lib/components/client.svelte
@@ -1,0 +1,518 @@
+<script lang='ts'>
+
+import { Duration } from 'google-protobuf/google/protobuf/duration_pb';
+import { onMount, onDestroy, createEventDispatcher } from 'svelte';
+import { type Credentials, ConnectionClosedError } from '@viamrobotics/rpc';
+import { Client, type ServiceError } from '@viamrobotics/sdk';
+import { notify } from '@viamrobotics/prime';
+import { client } from '@/stores/client';
+import { StreamManager } from '@/components/camera/stream-manager';
+import { getOperations, getResourceNames, getSessions } from '@/api/robot';
+import { useClient } from '@/hooks/use-client';
+import { getSensors } from '@/api/sensors';
+
+export let webrtcEnabled: boolean
+export let host: string
+export let signalingAddress: string
+export let bakedAuth: { authEntity?: string; creds?: Credentials; } = {};
+
+const { operations, sessions, resources, statusStream, rtt, connectionStatus } = useClient()
+
+const dispatch = createEventDispatcher<{
+  'connection-error': unknown
+}>()
+
+const relevantSubtypesForStatus = [
+  'arm',
+  'gantry',
+  'board',
+  'servo',
+  'motor',
+  'input_controller',
+] as const;
+
+const impliedURL = `${location.protocol}//${location.hostname}${location.port ? `:${location.port}` : ''}`;
+
+$client = new Client(impliedURL, {
+  enabled: webrtcEnabled,
+  host,
+  signalingAddress,
+  rtcConfig: {
+    iceServers: [
+      {
+        urls: 'stun:global.stun.twilio.com:3478',
+      },
+    ],
+  },
+
+  /*
+   * TODO(RSDK-3183): Opt out of reconnection management in the Typescript
+   * SDK because the Remote Control implements it's own reconnection management.
+   *
+   * The Typescript SDK only manages reconnections for WebRTC connections - once
+   * it can manage reconnections for direct gRPC connections, then we remove
+   * reconnection management from the Remote Control panel entirely and just rely
+   * on the Typescript SDK for that.
+   */
+  noReconnect: true,
+});
+
+let password = '';
+let lastStatusTS: number | null = null;
+let resourcesOnce = false;
+
+const streamManager = new StreamManager($client);
+
+const errors: Record<string, boolean> = {};
+
+const handleError = (message: string, error: unknown, onceKey: string) => {
+  if (onceKey) {
+    if (errors[onceKey]) {
+      return;
+    }
+
+    errors[onceKey] = true;
+  }
+
+  notify.danger(message);
+  console.error(message, { error });
+};
+
+const handleCallErrors = (list: { resources: boolean; ops: boolean }, newErrors: unknown) => {
+  const errorsList = document.createElement('ul');
+  errorsList.classList.add('list-disc', 'pl-4');
+
+  for (const key of Object.keys(list)) {
+    switch (key) {
+      case 'resources': {
+        errorsList.innerHTML += '<li>Robot Resources</li>';
+        break;
+      }
+      case 'ops': {
+        errorsList.innerHTML += '<li>Current Operations</li>';
+        break;
+      }
+    }
+  }
+
+  handleError(
+    `Error fetching the following: ${errorsList.outerHTML}`,
+    newErrors,
+    'connection'
+  );
+};
+
+const stringToResourceName = (nameStr: string) => {
+  const [prefix, suffix] = nameStr.split('/');
+  let name = '';
+
+  if (suffix) {
+    name = suffix;
+  }
+
+  const subtypeParts = prefix!.split(':');
+  if (subtypeParts.length > 3) {
+    throw new Error('more than 2 colons in resource name string');
+  }
+
+  if (subtypeParts.length < 3) {
+    throw new Error('less than 2 colons in resource name string');
+  }
+
+  return {
+    namespace: subtypeParts[0],
+    type: subtypeParts[1],
+    subtype: subtypeParts[2],
+    name,
+  };
+};
+
+const loadCurrentOps = async () => {
+  let now = Date.now();
+  const list = await getOperations();
+  const ops = [];
+
+  $rtt = Math.max(Date.now() - now, 0);
+
+  now = Date.now();
+
+  for (const op of list) {
+    ops.push({
+      op,
+      elapsed: op.started ? now - (op.started.seconds * 1000) : -1,
+    });
+  }
+
+  ops.sort((op1, op2) => {
+    if (op1.elapsed === -1 || op2.elapsed === -1) {
+      // move op with null start time to the back of the list
+      return op2.elapsed - op1.elapsed;
+    }
+    return op1.elapsed - op2.elapsed;
+  });
+
+  return ops;
+};
+
+const fetchCurrentSessions = async () => {
+  if (!sessionsSupported) {
+    return [];
+  }
+
+  try {
+    return await getSessions()
+  } catch (error) {
+    const serviceError = error as ServiceError
+    if (serviceError.code === serviceError.Code.Unimplemented) {
+      sessionsSupported = false;
+    }
+
+    return []
+  }
+};
+
+const updateStatus = (grpcStatuses: robotApi.Status[]) => {
+  for (const grpcStatus of grpcStatuses) {
+    const nameObj = grpcStatus.getName()!.toObject();
+    const statusJs = grpcStatus.getStatus()!.toJavaScript();
+    const name = resourceNameToString(nameObj);
+
+    $statuses[name] = statusJs;
+  }
+};
+
+const restartStatusStream = () => {
+  if ($statusStream) {
+    $statusStream.cancel();
+    $statusStream = null;
+  }
+
+  let newResources: commonApi.ResourceName.AsObject[] = [];
+
+  // get all relevant resources
+  for (const subtype of relevantSubtypesForStatus) {
+    newResources = [...newResources, ...filterSubtype($components, subtype)];
+  }
+
+  const names = newResources.map((name) => {
+    const resourceName = new commonApi.ResourceName();
+    resourceName.setNamespace(name.namespace);
+    resourceName.setType(name.type);
+    resourceName.setSubtype(name.subtype);
+    resourceName.setName(name.name);
+    return resourceName;
+  });
+
+  const streamReq = new robotApi.StreamStatusRequest();
+  streamReq.setResourceNamesList(names);
+  streamReq.setEvery(new Duration().setNanos(500_000_000));
+
+  $statusStream = $client.robotService.streamStatus(streamReq);
+  if ($statusStream !== null) {
+    $statusStream.on('data', (response: { getStatusList(): robotApi.Status[] }) => {
+      updateStatus(response.getStatusList());
+      lastStatusTS = Date.now();
+    });
+    $statusStream.on('status', (newStatus?: { details: unknown }) => {
+      if (!ConnectionClosedError.isError(newStatus!.details)) {
+        console.error('error streaming robot status', newStatus);
+      }
+      $statusStream = null;
+    });
+    $statusStream.on('end', () => {
+      console.error('done streaming robot status');
+      $statusStream = null;
+    });
+  }
+};
+
+
+// query metadata service every 0.5s
+const queryMetadata = async () => {
+  let resourcesChanged = false;
+  let shouldRestartStatusStream = !(resourcesOnce && $statusStream);
+
+  const resourcesList = await getResourceNames()
+
+  const differences: Set<string> = new Set(
+    $resources.map((name) => resourceNameToString(name))
+  );
+  const resourceSet: Set<string> = new Set(
+    resourcesList.map((name) => resourceNameToString(name))
+  );
+
+  for (const elem of resourceSet) {
+    if (differences.has(elem)) {
+      differences.delete(elem);
+    } else {
+      differences.add(elem);
+    }
+  }
+
+  if (differences.size > 0) {
+    resourcesChanged = true;
+
+    // restart status stream if resource difference includes a resource we care about
+    for (const elem of differences) {
+      const resource = stringToResourceName(elem);
+      if (
+        resource.namespace === 'rdk' &&
+        resource.type === 'component' &&
+        relevantSubtypesForStatus.includes(resource.subtype as typeof relevantSubtypesForStatus[number])
+      ) {
+        shouldRestartStatusStream = true;
+        break;
+      }
+    }
+  }
+
+  $resources = resourcesList;
+
+  resourcesOnce = true;
+  if (resourcesChanged === true) {
+    sensorNames = getSensors();
+  }
+
+  if (shouldRestartStatusStream === true) {
+    restartStatusStream();
+  }
+};
+
+const createAppConnectionManager = () => {
+  const checkIntervalMillis = 10_000;
+  const connections = {
+    resources: false,
+    ops: false,
+    sessions: false,
+  };
+
+  let timeout = -1;
+  let connectionRestablished = false;
+
+  const isConnected = () => {
+    return (
+      connections.resources &&
+      connections.ops &&
+      // check status on interval if direct grpc
+      (webrtcEnabled || (Date.now() - lastStatusTS! <= checkIntervalMillis))
+    );
+  };
+
+  const manageLoop = async () => {
+    try {
+      const newErrors: unknown[] = [];
+
+      try {
+        await queryMetadata();
+
+        if (!connections.resources) {
+          connectionRestablished = true;
+        }
+
+        connections.resources = true;
+      } catch (error) {
+        if (ConnectionClosedError.isError(error)) {
+          connections.resources = false;
+        } else {
+          newErrors.push(error);
+        }
+      }
+
+      if (connections.resources) {
+        try {
+          $operations = await loadCurrentOps();
+
+          if (!connections.ops) {
+            connectionRestablished = true;
+          }
+
+          connections.ops = true;
+        } catch (error) {
+          if (ConnectionClosedError.isError(error)) {
+            connections.ops = false;
+          } else {
+            newErrors.push(error);
+          }
+        }
+      }
+
+      if (connections.ops) {
+        try {
+          $sessions = await fetchCurrentSessions();
+
+          if (!connections.sessions) {
+            connectionRestablished = true;
+          }
+
+          connections.sessions = true;
+        } catch (error) {
+          if (ConnectionClosedError.isError(error)) {
+            connections.sessions = false;
+          } else {
+            newErrors.push(error);
+          }
+        }
+      }
+
+      if (isConnected()) {
+        if (connectionRestablished) {
+          notify.success('Connection established');
+          connectionRestablished = false;
+        }
+
+        errorMessage = '';
+        return;
+      }
+
+      if (newErrors.length > 0) {
+        handleCallErrors(connections, newErrors);
+      }
+      errorMessage = 'Connection lost, attempting to reconnect ...';
+
+      try {
+        console.debug('reconnecting');
+
+        // reset status/stream state
+        if ($statusStream) {
+          $statusStream.cancel();
+          $statusStream = null;
+        }
+        resourcesOnce = false;
+
+        await $client.connect();
+
+        const now = Date.now();
+        await getOperations();
+        
+        $rtt = Math.max(Date.now() - now, 0);
+
+        lastStatusTS = Date.now();
+        console.debug('reconnected');
+        streamManager.refreshStreams();
+      } catch (error) {
+        if (ConnectionClosedError.isError(error)) {
+          console.error('failed to reconnect; retrying');
+        } else {
+          console.error('failed to reconnect; retrying:', error);
+        }
+      }
+    } finally {
+      timeout = window.setTimeout(manageLoop, 500);
+    }
+  };
+
+  const stop = () => {
+    window.clearTimeout(timeout);
+    $statusStream?.cancel();
+    $statusStream = null;
+  };
+
+  const start = () => {
+    stop();
+    lastStatusTS = Date.now();
+    manageLoop();
+  };
+
+  return {
+    statuses: connections,
+    timeout,
+    stop,
+    start,
+    isConnected,
+  };
+};
+
+const appConnectionManager = createAppConnectionManager();
+
+const connect = async (creds?: Credentials) => {
+  $connectionStatus = 'connecting'
+
+  try {
+    await $client.connect(bakedAuth.authEntity, creds ?? bakedAuth.creds);
+  } catch (error) {
+    $connectionStatus = 'disconnected'
+    throw error;
+  }
+
+  $connectionStatus = 'connected'
+  connectedOnce = true;
+};
+
+const login = async (authType: string) => {
+  const creds = { type: authType, payload: password };
+
+  try {
+    await connect(creds);
+  } catch (error) {
+    notify.danger(`failed to connect: ${error.message}`);
+  }
+};
+
+const init = async () => {
+  if (supportedAuthTypes.length > 0) return
+
+  try {
+    await connect()
+  } catch (error) {
+    dispatch('connection-error', error)
+    setTimeout(init);
+  }
+}
+
+const handleUnload = () => {
+  appConnectionManager?.stop();
+  streamManager?.close();
+  $client.disconnect();
+};
+
+onMount(async () => {
+  window.addEventListener('beforeunload', handleUnload);
+});
+
+onDestroy(() => {
+  handleUnload();
+  window.removeEventListener('beforeunload', handleUnload);
+});
+
+init()
+
+</script>
+
+{#if $connectionStatus !== 'connected'}
+  <div>
+    {#each supportedAuthTypes as authType (authType)}
+      <div class="px-4 py-3">
+        <span>{authType}: </span>
+        <div class="w-96">
+          <input
+            bind:value={password}
+            disabled={$connectionStatus === 'connecting'}
+            class="
+              mb-2 block w-full appearance-none border p-2 text-gray-700
+              transition-colors duration-150 ease-in-out placeholder:text-gray-400 focus:outline-none
+            "
+            type="password"
+            autocomplete="off"
+            on:keyup={(event) => event.key === 'Enter' && login(authType)}
+          >
+          <v-button
+            disabled={$connectionStatus === 'connecting'}
+            label="Login"
+            on:click={$connectionStatus === 'connecting' ? undefined : () => login(authType)}
+          />
+        </div>
+      </div>
+    {/each}
+  </div>
+{/if}
+
+{#if $connectionStatus === 'connecting'}
+  <v-notify
+    variant='info'
+    title={`Connecting via ${webrtcEnabled ? 'WebRTC' : 'gRPC'}...`}
+  />
+{/if}
+
+{#if $connectionStatus === 'connected'}
+  <slot />
+{/if}
