+++ web/frontend/src/lib/components/collapse.svelte
